Casting to fix Windows port issues.

+++ test/cpp/test-crypto.cpp
@@ -687,18 +687,18 @@ void runBigDecimalTest(TestRunner& tr)
       long double d2 = -25.6;
       long double d12 = d1 / d2;
       char res[20];
-      sprintf(res, "%.10Lf", d12);
+      // casting to doubles due to windows issues
+      sprintf(res, "%.10f", (double)d12);
       assertStrCmp(res, "3.8554687500");
-      sprintf(res, "%.9Lf", d12);
+      sprintf(res, "%.9f", (double)d12);
       assertStrCmp(res, "3.855468750");
-      sprintf(res, "%.8Lf", d12);
+      sprintf(res, "%.8f", (double)d12);
       assertStrCmp(res, "3.85546875");
       // using "long double" will fail due to precision issues and rounding
       // value will be 3.85546874999.... vs the exact value of 3.85546875
       // the half even rounding rules will then round it down to 3.8554687
       // rather than the correct 3.8554688.  Casting to "double" will fix this. 
       tr.warning("long double precision issue workaround");
-      // sprintf(res, "%.7Lf", d12);
       sprintf(res, "%.7f", (double)d12);
       assertStrCmp(res, "3.8554688");
    }
